Install create_ap.resume in the proper location, /usr/lib/systemd/system-sleep
For unclear reasons, this only worked for me on Ubuntu 16.04 if it was put
in `/lib/systemd/system-sleep`, and NOT in `/usr/lib/systemd/system-sleep`.

I'm unsure if this was a systemd bug, a distro bug, or what… but it's
probably no longer applicable on modern distros.

Fix it, and leave behind a warning comment.

Thanks to @exuvo for bringing this up in
https://github.com/dlenski/create_ap/issues/1.